app/requirements.txt
asn1crypto==0.24.0
astroid==1.6.3
autopep8==1.3.5
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
codecov==2.1.13
colorama==0.3.9
coverage==4.5.1
cryptography==41.0.3
Django==2.2.28
django-extensions==2.0.7
django-ssl-auth==2.0.1
django-sslserver==0.20
django-widget-tweaks==1.4.2
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
ocspbuilder==0.10.2
oscrypto==0.19.1
pyasn1==0.4.2
pycodestyle==2.4.0
pycparser==2.18
pylint==1.8.4
pylint-django==0.11
pylint-plugin-utils==0.2.6
pyOpenSSL==23.2.0
pytz==2023.3
requests==2.31.0
# rope==0.11.0
rules==3.3
six==1.11.0
urllib3==1.26.5
wrapt==1.10.11
